requirements.txt
torch
torchvision
timm
streamlit
pillow
pandas
numpy
scikit-learn
scikit-image
opencv-python
plotly
